requirements.txt
groq==0.11.0
langchain==0.2.15
langchain-community==0.2.15
langchain-core==0.2.41
langchain-groq==0.1.9
langchain-openai==0.1.23
langchain-text-splitters==0.2.2
langsmith==0.1.112
langserve==0.2.2
openai==1.44.0
python-dotenv==1.0.1
streamlit==1.38.0
streamlit-js-eval==0.1.7
streamlit_pdf_viewer==0.0.18
sse_starlette==2.1.3
pydantic==1.10.13
fastapi==0.111.0